Update README.md
I think the sentence "For example, if `https://abc.com` is specified, any subdomain of arbitrary depth (`https://x.abc.com`, `https://y.x.abc.com`, etc) are indexed and searched." could use improvement. "Any subdomain of arbitrary depth" to me implies a restriction of subdomains to only those that are of arbitrary depth (which is meant to be interpreted as the depth does not matter). "Any subdomain of any depth" or "any subdomain, regardless of depth" would be better. Also, "subdomain" is singular but "are" is used. (any subdomain are indexed and searched).

I think a good solution to this sentence is to separate the explanation from the examples.